requirements.txt
langchain==0.3.27
openai==1.97.1
chromadb==1.0.15
ragas[eval]~=0.2.15
datasets==4.0.0
tiktoken==0.9.0
langgraph==0.5.4
beautifulsoup4==4.13.4
langchain-community==0.3.27
langchain-openai
gradio~=5.37.0
python-dotenv~=1.1.1